+++ execution/state/versionmap.go
@@ -499,8 +499,8 @@ type Version struct {
 
 var UnknownVersion = Version{TxIndex: UnknownDep, Incarnation: -1}
 
-func (v Version) blockAccessIndex() uint64 {
-	return uint64(v.TxIndex + 1)
+func (v Version) blockAccessIndex() uint32 {
+	return uint32(v.TxIndex + 1)
 }
 
 const (
